requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
fastapi==0.95.0
h11==0.14.0
hiredis==2.2.2
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
more-itertools==8.14.0
pptree==3.1
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.5.4
redis-om==0.1.2
sniffio==1.3.0
starlette==0.26.1
types-pyOpenSSL==23.1.0.2
types-redis==4.5.4.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
